rename: types > components 폴더 이름 오타 수정
