+++ src/components/common/Button/Button.styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { ButtonSize, ButtonType } from "@/types/compontents/button";
+import { ButtonSize, ButtonType } from "@/types/components/button";
 import {
 	getBorderRadius,
 	getColor,
